+++ src/Oscillator.cpp
@@ -52,8 +52,8 @@ unsigned Oscillator::getNoteFromFrequency(double frequency) {
  * @param note 
  * @return double 
  */
-double Oscillator::getFrequencyFromNote(unsigned note) {
-  return 27.5 * pow(2.0, (note - 21) / 12.0);
+double Oscillator::getFrequencyFromNote(double note) {
+  return 27.5 * pow(2.0, (note - 21.0) / 12.0);
 }
 
 double Oscillator::polyBLEP(double t)
@@ -86,7 +86,7 @@ double Oscillator::polyBLEP(double t)
  * 
  * @param note (midi format)
  */
-void Oscillator::setNote(unsigned note) {
+void Oscillator::setNote(double note) {
   setFrequency(getFrequencyFromNote(Util::checkParameterRange("Oscillator: Note", note, mNoteRange)));
 }
 
@@ -127,7 +127,7 @@ unsigned Oscillator::getNote() const {
   return getNoteFromFrequency(mFrequency);
 }
 
-Util::ParameterRange<unsigned> Oscillator::mNoteRange = std::make_pair(21, 108);
+Util::ParameterRange<double> Oscillator::mNoteRange = std::make_pair(21.0, 108.0);
 Util::ParameterRange<double> Oscillator::mFrequencyRange = std::make_pair(27.5, 4186.0);
 
 }
